Repository Reorga

+++ app/src/main/java/com/example/robin/angrynerds_wip/data/repository/filesystem/FileManager.java
@@ -4,11 +4,9 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Environment;
-import android.util.Log;
 
 import com.example.robin.angrynerds_wip.activities.note.note.data.PreviewImageCreator;
 import com.example.robin.angrynerds_wip.data.models.utils.Image;
-import com.example.robin.angrynerds_wip.data.repository.Repository;
 import com.example.robin.angrynerds_wip.data.repository.RepositoryConstants;
 import com.example.robin.angrynerds_wip.data.repository.database.DatabaseManager;
 
@@ -18,10 +16,14 @@ import java.io.IOException;
 
 public class FileManager {
 
-    private Context context;
+    private Context mContext;
 
     public FileManager() {
-        context = DatabaseManager.context;
+        mContext = DatabaseManager.context;
+    }
+
+    public Context getContext() {
+        return mContext;
     }
 
     public void saveImageToDirectory(Image pImage) throws IOException {
@@ -32,7 +34,7 @@ public class FileManager {
             String folderPath = Environment.DIRECTORY_PICTURES + "/" + folder;
             String imageName = pImage.getId() + ".jpg";
 
-            File storageDir = context.getExternalFilesDir(folderPath);
+            File storageDir = mContext.getExternalFilesDir(folderPath);
             FileOutputStream fos = null;
             Bitmap bitmap = pImage.getBitmap();
             File image = new File(storageDir, imageName);
@@ -43,7 +45,6 @@ public class FileManager {
                     if (folder.equals(RepositoryConstants.IMAGE_PREVIEW_FOLDER)) {
                         bitmap = previewImageCreator.getPreviewImage(bitmap);
                     }
-                    Log.i("NoteRemake", "Image " + pImage.getId() + " was saved!");
                     fos = new FileOutputStream(image);
 
                     bitmap.compress(Bitmap.CompressFormat.JPEG, 75, fos);
@@ -52,30 +53,23 @@ public class FileManager {
                 }
             }
         }
-
     }
 
     public Image readImageFromDirectory(Image image, String directory) {
-        String directoryPath = context.getExternalFilesDir(Environment.DIRECTORY_PICTURES).getAbsolutePath() + "/" + directory + "/";
+        String directoryPath = mContext.getExternalFilesDir(Environment.DIRECTORY_PICTURES).getAbsolutePath() + "/" + directory + "/";
         String filePath = directoryPath + image.getId() + ".jpg";
         Bitmap bitmap = BitmapFactory.decodeFile(filePath);
-        //Log.i("NoteRemake", "" + new File(filePath).exists());
         Image result = new Image(image.getId(), bitmap);
         return result;
     }
 
     public void deleteImageFromDirectory(String path) {
         File file = new File(path);
-        Log.i("NoteRemake", "Gelöschtes Bild existiert: " + file.exists());
-        file.delete();
+        if(file.exists()) file.delete();
     }
 
     public void deleteImageFromDirectory(Image image) {
         ImageDeleter imageDeleter = new ImageDeleter(this);
         imageDeleter.execute(image);
-    }
-
-    public Context getContext() {
-        return context;
     }
 }
